+++ armstrong/core/arm_access/tests/__init__.py
@@ -4,5 +4,5 @@ __path__ = extend_path(__path__, __name__)
 from .access_memberships import *
 from .fields import *
 from .subscription import *
-from .form_fields import *
+from .forms import *
 from .widgets import *
